Move SecurityConfiguration to OciSecurityConstants in Registry/Oci
Agent-Logs-Url: https://github.com/Azure/bicep/sessions/6b451bfe-3f4e-478a-8e81-2208dbaef92d
